Share the content ability test fixtures in a base test case
The users, ability registration, category set-up, and login helpers move
from ContentTest to Content_Ability_TestCase so the create, update, and
delete test cases can reuse them.
